backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-Mail
PyMySQL
python-dotenv
werkzeug
bleach
waitress